Stop blocking in wasi-provider
